fix: refactor interest module
